Pebble "Tracker" application
Simple time tracking program for up to five projects. You can edit the
project names and times using "Settings" on the phone. It's ok to run a
different application while a project is being tracked and it's ok to
assign time to multiple projects at once. Resetting to zero can be done
by holding down the middle button for 1 second. The program also tracks
the bluetooth connection and displays a bluetooth symbol (with a buzz)
when signal status changes.